# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "LINUX DO PAY",
        "contact": {},
        "version": "0.1.0"
    },
    "paths": {
        "/api/v1/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/merchant/api-keys": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/merchant.CreateAPIKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/merchant/api-keys/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "API Key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "API Key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/merchant.UpdateAPIKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchant"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "API Key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/oauth/callback": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/oauth.CallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/oauth/login": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/oauth/logout": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/oauth/user-info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
        "/api/v1/order/transactions": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/order.TransactionListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "merchant.CreateAPIKeyRequest": {
            "type": "object",
            "required": [
                "app_homepage_url",
                "app_name",
                "redirect_uri"
            ],
            "properties": {
                "app_description": {
                    "type": "string",
                    "maxLength": 100
                },
                "app_homepage_url": {
                    "type": "string",
                    "maxLength": 100
                },
                "app_name": {
                    "type": "string",
                    "maxLength": 20
                },
                "redirect_uri": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "merchant.UpdateAPIKeyRequest": {
            "type": "object",
            "properties": {
                "app_description": {
                    "type": "string",
                    "maxLength": 100
                },
                "app_homepage_url": {
                    "type": "string",
                    "maxLength": 100
                },
                "app_name": {
                    "type": "string",
                    "maxLength": 20
                },
                "redirect_uri": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "oauth.CallbackRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "order.TransactionListRequest": {
            "type": "object",
            "properties": {
                "endTime": {
                    "type": "string"
                },
                "page": {
                    "type": "integer",
                    "minimum": 1
                },
                "page_size": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                },
                "startTime": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "success",
                        "pending",
                        "failed",
                        "disputing",
                        "refund",
                        "refunded"
                    ]
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "receive",
                        "payment",
                        "transfer",
                        "community"
                    ]
                }
            }
        },
        "util.ResponseAny": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string",
                    "example": ""
                }
            }
        }
    }
}